Add gitlab-runner-register systemd service
This unit file registers runners with the gitlab service during startup.
Similarly, all runners are unregistered during shutdown.  This is
intended to clean up in the 30 second window a pre-empted instance is
given to clean up.  The intent is to clean up during the [Preemption
process][3].

Registration and unregistration of the gitlab-runner service moves from
the cloud-init commands into a systemd unit file so that the
gitlab-runner service can depend on the gitlab-runner-register servce.
Systemd units also execute very fast on container optimized OS, so the
shutdown event is processed within seconds, well within the 30 second
window provided by the platform.

References for this change:

[systemd EnvironmentFile][1] is used to secure the registration token to
be readable only by root in the container host.  This also keeps the
token out of the `systemctl status gitlab-runner-register` output.

[RemainAfterExit][2]
Use RemainAfterExit=yes becasue the nature of registration and
unregistration fits the description: "Use RemainAfterExit for services
which somehow change state of the system. When you want that state
reverted, you just stop the service."

[1]: https://coreos.com/os/docs/latest/using-environment-variables-in-systemd-units.html
[2]: https://stackoverflow.com/questions/38072849/when-should-the-option-remainafterexit-needs-to-be-set-true-when-creating-new-sy
[3]: https://cloud.google.com/compute/docs/instances/preemptible#preemption_process